Refactoring of gram: - creation of an abstract class gram identifying every general class representing a set of words which cannot be inserted as a composite of a document; - creation of implementing gram class, such as keyphrase for candidates and mention for coreferences; - every implementation is distinguished by a type that recall the implementation class name;
Mention class extends the gram class and has a list of new gram types, which are called references.

Add to blackboard a new data structure for grams, a map key-value where key is the type of gram and value is a map of grams for that type.
Modify all the methods using gram in blackboard and pointed as deprecated the ones using only keyphrase n-gram class.

Add porter stemmer annotator so to choose if using it or not in the pipeline.